test: dry data up
