+++ tests/specs/dry-run.js
@@ -36,7 +36,7 @@ const createDoneCallbackWrapper = () => {
   }
 }
 
-spec('The non-bump dry run task set', ({ test, end : endSpec }) => {
+spec('The dry run task set', ({ test, end : endSpec }) => {
   test('Has a changelog commit task that logs', ({ equal, end : endTest }) => {
     const changelogCommitTaskName = 'changelog:commit'
     const changelogCommitTask = tasks[changelogCommitTaskName]
@@ -93,6 +93,23 @@ spec('The non-bump dry run task set', ({ test, end : endSpec }) => {
     resetTaskLogger()    
     endTest()
   })
+
+  test('Has a major bump task that logs', ({ equal, end : endTest }) => {
+    const bumpMajorTaskName = 'bump:major:dry-run'
+    const bumpMajorTask = tasks[bumpMajorTaskName]
+    
+    equal(typeof bumpMajorTask, 'function', 'Has a major bump dry run task')
+
+    const resetTaskLoggerAndReturnLogMessage = taskLogSwapper()
+
+    bumpMajorTask()
+
+    setTimeout(() => {
+      equal(resetTaskLoggerAndReturnLogMessage().includes('Incrementing to the next "major" semantic version'), true)
+      endTest()
+    }, 200)
+  })
+
 
   endSpec()
 })
