+++ tests/specs/index.js
@@ -1,8 +1,6 @@
 const spec               = require('tape')
 const unleashTaskManager = require('../../')
 
-console.dir(unleashTaskManager)
-
 const SemVerIncrementTypes = [
   'major', 'minor', 'patch', 'prerelease'
 ]
